Cache lua installation and ruby gems

+++ lua/install.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 
-LUA_DIR="./tmp/lua51"
+LUA_DIR="$HOME/lua51"
+BUSTED="$LUA_DIR/bin/busted"
+
+if [ -f "$BUSTED" ]
+then
+  echo "$BUSTED already exists"
+  exit 0
+fi
+
 pip install hererocks
 hererocks $LUA_DIR -l5.1 -rlatest
 source $LUA_DIR/bin/activate
